services/strategy_registry/pyproject.toml
[project]
name = "strategy-registry"
version = "0.0.1"
requires-python = ">=3.12,<3.13"
dependencies = [
    "algobet-common",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/strategy_registry"]

[tool.uv.sources]
algobet-common = { workspace = true }
